+++ .github/workflows/code-cov.yml
@@ -1,0 +1,23 @@
+name: CodeCov
+on: [push, pull_request]
+
+jobs:
+  build:
+    runs-on: ubuntu-latest
+    name: Test python API
+    defaults:
+      run:
+        working-directory: ./api
+    steps:
+      - uses: actions/checkout@v1
+      - uses: actions/setup-python@v2
+        with:
+          python-version: "3.10"
+      - name: Install dependencies and project
+        run: |
+          python -m pip install poetry==1.2.1
+          poetry install
+      - name: Run tests and collect coverage
+        run: pytest --cov=./ --cov-report=xml
+      - name: Upload coverage reports to Codecov with GitHub Action
+        uses: codecov/codecov-action@v3
