+++ .github/workflows/testing.yml
@@ -39,4 +39,4 @@ jobs:
           poetry run pylint --rcfile=pyproject.toml ./feature_reviser ./tests
       - name: Test with pytest
         run: |
-          poetry run pytest
+          poetry run coverage run --source=feature_reviser -m pytest & coverage report -m
