--- docs/column_transformer-reference.md
@@ -1,1 +1,0 @@
-:::feature_reviser.transformer.column_transformer
