+++ docs/encoder_transformer-reference.md
@@ -1,0 +1,1 @@
+:::feature_reviser.transformer.encoder_transformer
