+++ docs/generic_transformer-reference.md
@@ -1,0 +1,1 @@
+:::feature_reviser.transformer.generic_transformer
