+++ docs/how_to_contribute.md
@@ -66,7 +66,7 @@ print(pipeline.fit_transform(df).head())
 1    Incredible Hulk   Schikaneder
 2      Tom and Jerry  Futuregarden
 ```
-For more examples check out our [`custom_transformer`](https://github.com/chrislemke/feature-reviser/blob/main/feature_reviser/transformer/custom_transformer.py) module.
+For more examples check out our [`transformers`](https://github.com/chrislemke/feature-reviser/blob/main/feature_reviser/transformer).
 
 ## Poetry
 We are using [Poetry](https://python-poetry.org/) to manage the dependencies and the virtual environment. If you have not used it before please check out the [documentation](https://python-poetry.org/docs/) to get started.
