+++ feature_reviser/transformer/string_transformer.py
@@ -5,303 +5,15 @@ import ipaddress
 import itertools
 import re
 import unicodedata
-from datetime import datetime
 from difflib import SequenceMatcher
-from typing import Any, Dict, List, Tuple, Union
+from typing import List, Tuple, Union
 
 import pandas as pd
 import phonenumbers
 from feature_engine.dataframe_checks import check_X
-from feature_engine.encoding import MeanEncoder as Me
 from sklearn.base import BaseEstimator, TransformerMixin
-from sklearn.utils.validation import check_is_fitted
 
 # pylint: disable=unused-argument, missing-function-docstring
-
-
-class ValueReplacerTransformer(BaseEstimator, TransformerMixin):
-    """
-    Uses Pandas `replace` method to replace values in a column.
-
-    Args:
-        features (List[Tuple[List[str], str, Any]]): List of tuples containing the column names as a list,
-            the value to replace (can be a regex), and the replacement value.
-    """
-
-    def __init__(self, features: List[Tuple[List[str], str, Any]]) -> None:
-        self.features = features
-
-    def fit(self, X=None, y=None) -> "ValueReplacerTransformer":  # type: ignore
-        return self
-
-    def transform(self, X: pd.DataFrame) -> pd.DataFrame:
-        """
-        Args:
-            X (pd.DataFrame): Dataframe containing the columns where values should be replaced.
-
-        Returns:
-            pd.DataFrame: Dataframe with replaced values.
-        """
-
-        X = check_X(X)
-
-        for (columns, value, replacement) in self.features:
-            for column in columns:
-                print(column)
-                is_regex = ValueReplacerTransformer.__check_for_regex(value)
-                column_dtype = X[column].dtype
-
-                if column_dtype is not str and is_regex:
-                    X[column] = X[column].astype(str)
-
-                X[column] = X[column].replace(value, replacement, regex=True)
-
-                if X[column].dtype != column_dtype:
-                    X[column] = X[column].astype(column_dtype)
-
-        return X
-
-    @staticmethod
-    def __check_for_regex(string: str) -> bool:
-        if not isinstance(string, str):
-            return False
-        try:
-            re.compile(string)
-            is_valid = True
-        except re.error:  # pylint: disable=W0702
-            is_valid = False
-        return is_valid
-
-
-class ColumnDropperTransformer(BaseEstimator, TransformerMixin):
-    """
-    Drops columns from a dataframe using Pandas `drop` method.
-
-    Args:
-        columns (Union[str, List[str]]): Columns to drop. Either a single column name or a list of column names.
-    """
-
-    def __init__(self, columns: Union[str, List[str]]) -> None:
-        self.columns = columns
-
-    def fit(self, X=None, y=None) -> "ColumnDropperTransformer":  # type: ignore
-        return self
-
-    def transform(self, X: pd.DataFrame) -> pd.DataFrame:
-        """
-        Returns the dataframe with the columns dropped.
-
-        Args:
-            X (pd.DataFrame): Dataframe to drop columns from.
-
-        Returns:
-            pd.DataFrame: Dataframe with columns dropped.
-        """
-        X = check_X(X)
-        return X.drop(self.columns, axis=1)
-
-
-class DurationCalculatorTransformer(BaseEstimator, TransformerMixin):
-    """
-    Calculates the duration between to given dates.
-
-    Args:
-        features (Tuple[str, str]): The two columns that contain the dates which should be used to calculate the duration.
-        unit (str): The unit in which the duration should be returned. Should be either `days` or `seconds`.
-        new_column_name (str): The name of the output column.
-    """
-
-    def __init__(
-        self, features: Tuple[str, str], unit: str, new_column_name: str
-    ) -> None:
-
-        if unit not in ["days", "seconds"]:
-            raise ValueError("Unsupported unit. Should be either `days` or `seconds`!")
-
-        self.features = features
-        self.unit = unit
-        self.new_column_name = new_column_name
-
-    def fit(self, X=None, y=None) -> "DurationCalculatorTransformer":  # type: ignore
-        return self
-
-    def transform(self, X: pd.DataFrame) -> pd.DataFrame:
-        """
-        Transform method that calculates the duration between two dates.
-
-        Args:
-            X (pandas.DataFrame): The input DataFrame.
-
-        Returns:
-            pandas.DataFrame: The transformed DataFrame.
-        """
-
-        if not all(elem in X.columns for elem in self.features):
-            raise ValueError("Not all provided `features` could be found in `X`!")
-
-        X = check_X(X)
-
-        if X.shape[1] != 2:
-            raise ValueError(
-                f"Only two columns should be provided! But {X.shape[1]} were given."
-            )
-
-        duration_series = pd.to_datetime(
-            X[self.features[1]], utc=True, errors="raise"
-        ) - pd.to_datetime(X[self.features[0]], utc=True, errors="raise")
-
-        X[self.new_column_name] = (
-            duration_series.dt.days
-            if self.unit == "days"
-            else duration_series.dt.total_seconds()
-        )
-        return X
-
-
-class NaNTransformer(BaseEstimator, TransformerMixin):
-    """
-    Replace NaN values with a specified value.
-
-    Args:
-        X (pandas.DataFrame): Dataframe to transform.
-        values (Dict[str, Any]): Dictionary with column names as keys and values to replace NaN with as values.
-    """
-
-    def __init__(self, values: Dict[str, Any]) -> None:
-        self.values = values
-
-    def fit(self, X=None, y=None) -> "NaNTransformer":  # type: ignore
-        return self
-
-    def transform(self, X: pd.DataFrame) -> pd.DataFrame:
-        """
-        Replace NaN values with a specified value.
-
-        Args:
-            X (pandas.DataFrame): Dataframe to transform.
-
-        Returns:
-            pandas.DataFrame: Transformed dataframe.
-        """
-        X = check_X(X)
-        return X.fillna(self.values)
-
-
-class TimestampTransformer(BaseEstimator, TransformerMixin):
-    """
-    Transforms a date column with a specified format into a timestamp column.
-
-    Args:
-        features (List[str]): List of features which should be transformed.
-        format (str): Format of the date column. Defaults to "%Y-%m-%d".
-    """
-
-    def __init__(
-        self,
-        features: List[str],
-        date_format: str = "%Y-%m-%d",
-    ) -> None:
-        self.features = features
-        self.date_format = date_format
-
-    def fit(self, X=None, y=None) -> "TimestampTransformer":  # type: ignore
-        return self
-
-    def transform(self, X: pd.DataFrame) -> pd.DataFrame:
-        """
-        Transforms columns from the provided dataframe.
-
-        Args:
-            X (pandas.DataFrame): Dataframe with columns to transform.
-
-        Returns:
-            pandas.DataFrame: Dataframe with transformed columns.
-        """
-
-        if not all(elem in X.columns for elem in self.features):
-            raise ValueError("Not all provided `features` could be found in `X`!")
-
-        X = check_X(X)
-        for column in self.features:
-            X[column] = pd.to_datetime(
-                X[column], format=self.date_format, errors="raise"
-            )
-            X[column] = (X[column] - datetime(1970, 1, 1)).dt.total_seconds()
-        return X
-
-
-class QueryTransformer(BaseEstimator, TransformerMixin):
-    """
-    Applies a list of queries to a dataframe.
-            If it operates on a dataset used for supervised learning this transformer should
-            be applied on the dataframe containing `X` and `y`. So removing of columns by queries
-            also removes the corresponding `y` value.
-    Read more about queries [here](https://pandas.pydata.org/docs/reference/api/pandas.DataFrame.query.html#pandas.DataFrame.query).
-
-    Args:
-        queries (List[str]): List of queries to apply to the dataframe.
-    """
-
-    def __init__(self, queries: List[str]) -> None:
-        self.queries = queries
-
-    def fit(self, X=None, y=None) -> "QueryTransformer":  # type: ignore
-        return self
-
-    def transform(self, Xy: pd.DataFrame) -> pd.DataFrame:
-        """
-        Applies the list of queries to the dataframe.
-
-        Args:
-            Xy (pd.DataFrame): Dataframe to apply the queries to.
-
-        Returns:
-            pd.DataFrame: Dataframe with the queries applied.
-        """
-
-        Xy = check_X(Xy)
-        for query in self.queries:
-            Xy = Xy.query(query)
-        return Xy
-
-
-class MeanEncoder(BaseEstimator, TransformerMixin):
-    """
-    Scikit-learn API for the [feature-engine MeanEncoder](https://feature-engine.readthedocs.io/en/latest/api_doc/encoding/MeanEncoder.html).
-
-    Args:
-        fill_na_value (Union[int, float]): Value to fill NaN values with.
-            Those may appear if a category is not present in the set the encoder was not fitted on.
-    """
-
-    def __init__(self, fill_na_value: Union[int, float] = -999) -> None:
-        self.encoder = Me(ignore_format=False)
-        self.fill_na_value = fill_na_value
-
-    def fit(self, X: pd.DataFrame, y: pd.Series) -> "MeanEncoder":
-        """
-        Fit the MeanEncoder to the data.
-        Args:
-            X (pandas.DataFrame): DataFrame to fit the MeanEncoder to.
-            y (pandas.Series): Target variable.
-
-        Returns:
-            MeanEncoder: Fitted MeanEncoder.
-        """
-        self.encoder.fit(X, y)
-        return self
-
-    def transform(self, X: pd.DataFrame) -> pd.DataFrame:
-        """
-        Transform the data using the fitted MeanEncoder.
-        Args:
-            X (pandas.DataFrame): DataFrame to transform.
-
-        Returns:
-            pandas.DataFrame: Transformed data.
-        """
-        check_is_fitted(self)
-        return self.encoder.transform(X).fillna(self.fill_na_value)
 
 
 class IPAddressEncoderTransformer(BaseEstimator, TransformerMixin):
