+++ pyproject.toml
@@ -26,26 +26,27 @@ keywords = [
 
 [tool.poetry.dependencies]
 python = ">=3.8,<3.11"
-scikit-learn = "^1.1.2"
 joblib = "^1.2.0"
-pandas = "^1.5.0"
 numpy = "^1.23.3"
+scikit-learn = "^1.1.2"
+pandas = "^1.5.0"
 sklearn-pandas = "^2.2.0"
 feature-engine = "^1.4.1"
-phonenumbers = "^8.12.55"
+phonenumbers = "^8.12.56"
 
 
 [tool.poetry.dev-dependencies]
-black = {extras = ["jupyter"], version = "^22.3.0"}
+bandit = "^1.7.4"
+black = "^22.3.0"
+coverage = "^6.4.4"
 isort = "^5.10.1"
-pylint = "^2.13.9"
+mkdocs = "^1.3.1"
+mkdocstrings = "^0.18.1"
+mkdocs-material = "^8.5.2"
 mypy = "^0.971"
 pre-commit = "^2.19.0"
-mkdocs = "^1.3.1"
-mkdocs-material = "^8.5.2"
-mkdocstrings = "^0.18.1"
+pylint = "^2.13.9"
 pytest = "^7.1.3"
-bandit = "^1.7.4"
 
 [build-system]
 requires = ["poetry-core>=1.0.0"]
